Improvements to collision and hitboxes
Wall collision is kind of done. It can calculate collision, but in due
to having to work with multiple coordinate systems, I was not able to
make it work properly, creating invisible walls. Calculations will be
skipped, but can be turned on for showing purposes by commenting the
first line in the checkCollisionWithCubes method in CCamera.
Hitboxes have been resized and repositioned to better fit the enemies.
They are a bit smaller than the models, but are much easier to hit, and
are on the actual model.
Furthermore, Trap is now its own type (was for testing purposes).